Clippy: 160 -> 20, and the 20 that remain are one decision
Fixed, in descending order of whether it could bite:

* `hinf.rs` guarded gamma with `!(gamma > 0.0)`. Clippy's neg_cmp_op_on_partial_ord
  suggestion here is `gamma <= 0.0`, which would OPEN A HOLE: every comparison
  with NaN is false, so `NaN <= 0.0` passes the guard while `!(NaN > 0.0)`
  correctly rejects it. Rewritten as an explicit `!gamma.is_finite() ||
  gamma <= 0.0` so the NaN case is stated rather than riding on double negation.
* Nine `} if` boundaries on one line (`if c { crash += 1; } if d { death += 1; }`)
  read as a chain with a missing `else`. They are independent counters, so `else`
  would be wrong; split onto separate lines, which is what the lint is for.
* Eleven wasm labs had the `new()` constructor without the `impl Default` half of
  the house convention. Added — and then MOVED, because appending them put them
  after `#[cfg(test)] mod tests` and tripped items_after_test_module. That is the
  same insertion trap this repo has hit before; the fix belongs above the tests.
* `ferromotion-fluid` had an orphaned doc comment: `/// Linear interpolation of a
  monotone-in-x (x, y) table` sat 137 lines above the `interp` it documents, so it
  was documenting the verification module instead. Reattached.
* Two hex literals regrouped (value checked identical), two redundant parens.
* `cargo clippy --fix` handled 18 more: `.into()` on an already-correct type,
  `(x.ln() * -1.0)` -> `-x.ln()`, a `let` with a deferred single assignment, and
  one function moved out from under a test module.

The four categories now allowed workspace-wide are allowed WITH REASONS in
Cargo.toml, because each asks for a rewrite that makes this code worse:
needless_range_loop (50) on dynamics/RNEA/matrix loops that index several arrays
by the same variable, where an iterator rewrite hides which index is which;
too_many_arguments (11) on physics calls that genuinely take robot, inertia, q,
qd, qdd and gravity; type_complexity (12) where the tuple shape is the useful
part at the call site; should_implement_trait (3) on small generic fixed-size
helpers in `gendyn`.

What is left is 20 sites of `clippy::ptr_arg` — `&mut Vec` where `&mut [_]` would
do. Left VISIBLE on purpose: it is a real improvement, it changes public
signatures, and so it is one semver decision rather than a cleanup.

1259 workspace tests, 0 failures, 0 build errors.

+++ crates/ferromotion-core/examples/so101_arbiter.rs
@@ -174,9 +174,11 @@ fn main() {
         let (mut crash, mut death, mut rec, mut n) = (0, 0, 0, 0);
         for (ei, env) in envs.iter().enumerate() {
             for ep in 0..14 {
-                let seed = (ei as u32 * 131 + ep * 7 + 1) as u32;
+                let seed = ei as u32 * 131 + ep * 7 + 1;
                 let (c, d, r) = episode(&robot, &inertia, &task, arb, env, seed);
-                n += 1; if c { crash += 1; } if d { death += 1; } if r { rec += 1; }
+                n += 1; if c { crash += 1; }
+                if d { death += 1; }
+                if r { rec += 1; }
             }
         }
         println!("  {label}");
